# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  api_gateway_dto.GetSupplierByIDResponse:
    properties:
      business_address:
        type: string
      company_name:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      documents:
        items:
          $ref: '#/definitions/api_gateway_dto.GetSupplierDocument'
        type: array
      id:
        type: integer
      logo_thumbnail_url:
        type: string
      status:
        $ref: '#/definitions/common.SupplierProfileStatus'
      tax_id:
        type: string
      updated_at:
        type: string
    type: object
  api_gateway_dto.GetSupplierByIDResponseDocs:
    properties:
      data:
        $ref: '#/definitions/api_gateway_dto.GetSupplierByIDResponse'
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
  api_gateway_dto.GetSupplierDocument:
    properties:
      admin_note:
        type: string
      created_at:
        type: string
      document:
        $ref: '#/definitions/api_gateway_dto.SupplierDocument'
      id:
        type: string
      updated_at:
        type: string
      verification_status:
        $ref: '#/definitions/common.SupplierDocumentStatus'
    type: object
  api_gateway_dto.GetSupplierOrdersResponse:
    properties:
      actual_delivery_date:
        type: string
      cancelled_reason:
        type: string
      discount_amount:
        description: discount amount
        type: number
      estimated_delivery_date:
        type: string
      notes:
        type: string
      order_item_id:
        type: string
      product_id:
        description: info products
        type: string
      product_name:
        type: string
      product_variant_id:
        type: string
      product_variant_name:
        type: string
      product_variant_thumbnail:
        type: string
      quantity:
        type: integer
      recipient_name:
        type: string
      recipient_phone:
        type: string
      shipping_address:
        type: string
      shipping_fee:
        type: number
      shipping_method:
        $ref: '#/definitions/common.MethodType'
      status:
        $ref: '#/definitions/common.StatusOrder'
      tax_amount:
        type: number
      total_price:
        description: money need to be paid
        type: number
      tracking_number:
        description: Used for detail when click into one order item
        type: string
      unit_price:
        type: number
    type: object
  api_gateway_dto.GetSupplierOrdersResponseDocs:
    properties:
      data:
        items:
          $ref: '#/definitions/api_gateway_dto.GetSupplierOrdersResponse'
        type: array
      metadata:
        $ref: '#/definitions/api_gateway_dto.MetadataWithPagination'
    type: object
...
  api_gateway_dto.GetSuppliersResponse:
    properties:
      business_address:
        type: string
      company_name:
        type: string
      contact_phone:
        type: string
      created_at:
        type: string
      id:
        type: integer
      logo_thumbnail_url:
        type: string
      status:
        $ref: '#/definitions/common.SupplierProfileStatus'
      tax_id:
        type: string
      updated_at:
        type: string
    type: object
  api_gateway_dto.GetSuppliersResponseDocs:
    properties:
      data:
        items:
          $ref: '#/definitions/api_gateway_dto.GetSuppliersResponse'
        type: array
      metadata:
        $ref: '#/definitions/api_gateway_dto.MetadataWithPagination'
    type: object
...
  api_gateway_dto.Metadata:
    properties:
      code:
        type: integer
    type: object
...
  api_gateway_dto.MetadataWithPagination:
    properties:
      code:
        type: integer
      pagination:
        $ref: '#/definitions/api_gateway_dto.Pagination'
    type: object
...
  api_gateway_dto.Pagination:
    properties:
      has_next:
        type: boolean
      has_previous:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  api_gateway_dto.RegisterSupplierRequest:
    properties:
      business_address_id:
        type: integer
      company_name:
        type: string
      contact_phone:
        type: string
      description:
        type: string
      documents:
        $ref: '#/definitions/api_gateway_dto.SupplierDocument'
      logo_company_url:
        type: string
      tax_id:
        type: string
    required:
    - business_address_id
    - company_name
    - contact_phone
    - documents
    - logo_company_url
    - tax_id
    type: object
  api_gateway_dto.RegisterSupplierResponse:
    type: object
  api_gateway_dto.RegisterSupplierResponseDocs:
    properties:
      data:
        $ref: '#/definitions/api_gateway_dto.RegisterSupplierResponse'
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
...
  api_gateway_dto.ResponseErrorDocs:
    properties:
      error: {}
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
...
  api_gateway_dto.SupplierDocument:
    properties:
      business_license:
        type: string
      id_card_back:
        type: string
      id_card_front:
        type: string
      tax_certificate:
        type: string
    required:
    - business_license
    - id_card_back
    - id_card_front
    - tax_certificate
    type: object
...
  api_gateway_dto.UpdateOrderItemRequest:
    properties:
      status:
        allOf:
        - $ref: '#/definitions/common.StatusOrder'
        enum:
        - confirmed
        - cancelled
    type: object
  api_gateway_dto.UpdateOrderItemResponse:
    type: object
  api_gateway_dto.UpdateOrderItemResponseDocs:
    properties:
      data:
        $ref: '#/definitions/api_gateway_dto.UpdateOrderItemResponse'
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
...
  api_gateway_dto.UpdateRoleForUserRegisterSupplierRequest:
    properties:
      user_id:
        type: integer
    required:
    - user_id
    type: object
  api_gateway_dto.UpdateRoleForUserRegisterSupplierResponse:
    type: object
  api_gateway_dto.UpdateRoleForUserRegisterSupplierResponseDocs:
    properties:
      data:
        $ref: '#/definitions/api_gateway_dto.UpdateRoleForUserRegisterSupplierResponse'
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
...
  api_gateway_dto.UpdateSupplierDocumentVerificationStatusRequest:
    properties:
      status:
        $ref: '#/definitions/common.SupplierDocumentStatus'
    type: object
  api_gateway_dto.UpdateSupplierDocumentVerificationStatusResponse:
    properties:
      status:
        $ref: '#/definitions/common.SupplierDocumentStatus'
    type: object
  api_gateway_dto.UpdateSupplierDocumentVerificationStatusResponseDocs:
    properties:
      data:
        $ref: '#/definitions/api_gateway_dto.UpdateSupplierDocumentVerificationStatusResponse'
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
  api_gateway_dto.UpdateSupplierRequest:
    properties:
      status:
        $ref: '#/definitions/common.SupplierProfileStatus'
    type: object
  api_gateway_dto.UpdateSupplierResponse:
    type: object
  api_gateway_dto.UpdateSupplierResponseDocs:
    properties:
      data:
        $ref: '#/definitions/api_gateway_dto.UpdateSupplierResponse'
      metadata:
        $ref: '#/definitions/api_gateway_dto.Metadata'
    type: object
...
  common.MethodType:
    enum:
    - momo
    - cod
    type: string
    x-enum-varnames:
    - Momo
    - Cod
  common.StatusOrder:
    enum:
    - pending_payment
    - pending
    - confirmed
    - processing
    - ready_to_ship
    - in_transit
    - out_for_delivery
    - delivered
    - cancelled
    - payment_failed
    - refunded
    type: string
    x-enum-comments:
      Confirmed: Supplier đã xác nhận
      Delivered: Đã giao thành công
      InTransit: Đang vận chuyển (đang ship)
      OutForDelivery: Sắp giao (shipper đang trên đường)
      Pending: Chờ supplier xác nhận
      PendingPayment: Chờ thanh toán
      Processing: Đang chuẩn bị hàng
      ReadyToShip: Sẵn sàng giao hàng
    x-enum-varnames:
    - PendingPayment
    - Pending
    - Confirmed
    - Processing
    - ReadyToShip
    - InTransit
    - OutForDelivery
    - Delivered
    - Cancelled
    - PaymentFailed
    - Refunded
  common.SupplierDocumentStatus:
    enum:
    - pending
    - approved
    - rejected
    type: string
    x-enum-varnames:
    - SupplierDocumentStatusPending
    - SupplierDocumentStatusApproved
    - SupplierDocumentStatusRejected
  common.SupplierProfileStatus:
    enum:
    - pending
    - active
    - suspended
    type: string
    x-enum-varnames:
    - SupplierProfileStatusPending
    - SupplierProfileStatusActive
    - SupplierProfileStatusSuspended
...
host: server.local:3000
info:
  contact: {}
  description: API for ecommerce
  title: Ecommerce API
  version: "1.0"
paths:
...
  /suppliers:
    get:
      consumes:
      - application/json
      description: get suppliers by admin
      parameters:
      - in: query
        name: company_name
        type: string
      - in: query
        name: contact_phone
        type: string
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 1
        name: page
        type: integer
      - enum:
        - pending
        - active
        - suspended
        in: query
        name: status
        type: string
        x-enum-varnames:
        - SupplierProfileStatusPending
        - SupplierProfileStatusActive
        - SupplierProfileStatusSuspended
      - in: query
        name: tax_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.GetSuppliersResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      security:
      - BearerAuth: []
      summary: get suppliers by admin
      tags:
      - suppliers
  /suppliers/{id}:
    get:
      consumes:
      - application/json
      description: get supplier detail
      parameters:
      - in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.GetSupplierByIDResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      security:
      - BearerAuth: []
      summary: get supplier detail
      tags:
      - suppliers
    patch:
      consumes:
      - application/json
      description: Quản trị viên cho phép supplier hoạt động hoặc không
      parameters:
      - description: Thông tin cần cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_gateway_dto.UpdateSupplierRequest'
      - in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.UpdateSupplierResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      security:
      - BearerAuth: []
      summary: Quản trị viên cho phép supplier hoạt động hoặc không
      tags:
      - suppliers
  /suppliers/{id}/documents/{documentID}:
    patch:
      consumes:
      - application/json
      description: approve or reject supplier document
      parameters:
      - description: Thông tin cần cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_gateway_dto.UpdateSupplierDocumentVerificationStatusRequest'
      - in: path
        name: documentID
        required: true
        type: string
      - in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.UpdateSupplierDocumentVerificationStatusResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      security:
      - BearerAuth: []
      summary: approve or reject supplier document
      tags:
      - suppliers
  /suppliers/me:
    get:
      consumes:
      - application/json
      description: get supplier orders
      parameters:
      - in: query
        minimum: 1
        name: limit
        type: integer
      - in: query
        minimum: 1
        name: page
        type: integer
      - enum:
        - pending_payment
        - pending
        - confirmed
        - processing
        - ready_to_ship
        - in_transit
        - out_for_delivery
        - delivered
        - cancelled
        - payment_failed
        - refunded
        in: query
        name: status
        type: string
        x-enum-comments:
          Confirmed: Supplier đã xác nhận
          Delivered: Đã giao thành công
          InTransit: Đang vận chuyển (đang ship)
          OutForDelivery: Sắp giao (shipper đang trên đường)
          Pending: Chờ supplier xác nhận
          PendingPayment: Chờ thanh toán
          Processing: Đang chuẩn bị hàng
          ReadyToShip: Sẵn sàng giao hàng
        x-enum-varnames:
        - PendingPayment
        - Pending
        - Confirmed
        - Processing
        - ReadyToShip
        - InTransit
        - OutForDelivery
        - Delivered
        - Cancelled
        - PaymentFailed
        - Refunded
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.GetSupplierOrdersResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      summary: get supplier orders
      tags:
      - suppliers
  /suppliers/orders/{orderItemID}:
    post:
      consumes:
      - application/json
      description: update order item
      parameters:
      - in: path
        name: orderItemID
        required: true
        type: string
      - description: Thông tin cần cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_gateway_dto.UpdateOrderItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.UpdateOrderItemResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      summary: update order item
      tags:
      - suppliers
  /suppliers/register:
    post:
      consumes:
      - application/json
      description: customer register supplier
      parameters:
      - description: data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api_gateway_dto.RegisterSupplierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.RegisterSupplierResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      security:
      - BearerAuth: []
      summary: customer register supplier
      tags:
      - suppliers
  /suppliers/uprole:
    post:
      consumes:
      - application/json
      description: up role supplier for user
      parameters:
      - description: Thông tin cần cập nhật
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_gateway_dto.UpdateRoleForUserRegisterSupplierRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_gateway_dto.UpdateRoleForUserRegisterSupplierResponseDocs'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api_gateway_dto.ResponseErrorDocs'
      summary: up role supplier for user
      tags:
      - suppliers
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"